refactor(price): 가격 int→BigDecimal 전면 전환 (미국대전환 P2)
USD 센트 유실 방지. BrokerPort.buy/sell price·getCurrentPrice·
KoreaInvestmentPriceResponse·NotificationEvent를 BigDecimal화하고
TradingStrategyService의 price.intValue() 절삭 제거. KRW 정수호가 변환은
KIS 어댑터 executeOrder 내부(setScale HALF_UP)에만 격리 → 포트 계약 무손실.
KRW 동작 바이트 동일(정수=scale0), 수량은 int 유지(notional=P6).
clean test 143 그린(USD 무손실 실증 2 추가).

+++ src/main/java/com/canagent/port/BrokerPort.java
@@ -4,21 +4,26 @@ import com.canagent.service.dto.KoreaInvestmentBalanceResponse;
 import com.canagent.service.dto.KoreaInvestmentOrderResponse;
 import com.canagent.service.dto.KoreaInvestmentPriceResponse;
 
+import java.math.BigDecimal;
+
 /**
  * 브로커(주문·잔고·현재가) 경계 포트.
  *
  * <p>P1(미국 대전환): 어댑터 이음새 확보를 위한 순수 인터페이스 추출.
- * 현재 유일한 구현은 {@code KoreaInvestmentApiClient}이며, 시그니처·반환 DTO는 기존 KIS 동작 그대로다.
- * 미국용 재설계(placeBuy/OrderSpec, BigDecimal 가격 등)는 P2/P6 소관이므로 여기서 바꾸지 않는다.
+ * 현재 유일한 구현은 {@code KoreaInvestmentApiClient}이다.
+ *
+ * <p>P2(미국 대전환): 가격은 {@code BigDecimal}로 무손실 전달한다(USD 센트 유실 방지).
+ * KRW 정수 호가는 KIS 구현체 내부에서만 정수로 변환하며, 포트 계약은 소수 가격을 그대로 전달한다.
+ * 수량(quantity)은 이번 단계에서 정수 유지(소수 수량·notional 사이징은 P6). 주문모델 재설계(placeBuy/OrderSpec)도 P6.
  *
  * <p>호출부: TradingStrategyService(buy/sell/getBalance), IntradayMonitorWorker(getCurrentPrice/getBalance),
  * PortfolioScheduler(getCurrentPrice), DashboardController(getBalance).
  */
 public interface BrokerPort {
 
-    KoreaInvestmentOrderResponse buy(String stockCode, int quantity, int price);
+    KoreaInvestmentOrderResponse buy(String stockCode, int quantity, BigDecimal price);
 
-    KoreaInvestmentOrderResponse sell(String stockCode, int quantity, int price);
+    KoreaInvestmentOrderResponse sell(String stockCode, int quantity, BigDecimal price);
 
     KoreaInvestmentBalanceResponse getBalance();
